Update 5_Distributed_preprocessing.ipynb
Two small changes in the notebook.

First, when defining the working directory the else should be deleted. Because if the directory does not exist it will be created but cfg.PATHS['working_dir'] will not be set.

Second, I included print_log=False to suppress the error messages (which flood the output cell).